requirements.txt
transformers==4.30.2
datasets==2.13.1
torch==2.0.1
accelerate==0.21.0
trl==0.4.7
safetensors==0.3.1
gym==0.26.2
stable-baselines3==2.0.0
google-cloud-storage==2.10.0
wandb==0.15.5
nltk==3.8.1